requirements.txt
ultralytics>=8.2.0
numpy>=1.26.0
PyYAML>=6.0.0
matplotlib>=3.8.0
